% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_line_data}
\alias{get_line_data}
\title{Line data}
\usage{
get_line_data(line)
}
\arguments{
\item{line}{Scalar integer specifying the line.}
}
\value{
A \code{data.frame} with variables \code{line}, \code{date}, \code{weekday}, \code{hour},
\code{scheduled_departure}, \code{actual_departure}, \code{scheduled_arrival},
\code{actual_arrival}.
}
\description{
Returns departure and arrival data for a given \code{line}.
}
\examples{
data <- get_line_data(11)
head(data)
summary(data)

}
