% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_delays_by_weekday_hour}
\alias{plot_delays_by_weekday_hour}
\title{Weekday-hour delay barplot}
\usage{
plot_delays_by_weekday_hour(data)
}
\arguments{
\item{data}{Arrival and departure data (see \code{\link[=get_line_data]{get_line_data()}}).}
}
\description{
Create a barplot of normalized delay counts by hour for each weekday.
}
\examples{
data <- get_line_data(11)
plot_delays_by_weekday_hour(data)

}
