% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_line_report}
\alias{render_line_report}
\title{Line report}
\usage{
render_line_report(line, output_dir = ".", ...)
}
\arguments{
\item{line}{Scalar integer specifying the line.}

\item{output_dir}{The output directory for the rendered report.}

\item{...}{Additional arguments to be passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
The path to the rendered report.
}
\description{
Render the delay analysis HTML report for a given \code{line}.
}
\examples{
report <- render_line_report(line = 11, tempdir())
browseURL(report)

}
